push through Proof Protocol curve hierarchy integration
