+++ symbolic-base/src/ZkFold/Base/Algebra/EllipticCurve/BN254.hs
@@ -1,6 +1,7 @@
-{-# LANGUAGE DerivingVia     #-}
-{-# LANGUAGE OverloadedLists #-}
-
+{-# LANGUAGE DerivingVia          #-}
+{-# LANGUAGE OverloadedLists      #-}
+{-# LANGUAGE TypeOperators        #-}
+{-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module ZkFold.Base.Algebra.EllipticCurve.BN254
@@ -11,16 +12,17 @@ module ZkFold.Base.Algebra.EllipticCurve.BN254
   , Fp2
   , Fp6
   , Fp12
-  , BN254_G1
-  , BN254_G2
-  , BN254_GT) where
+  , BN254_G1_Point
+  , BN254_G2_Point
+  , BN254_GT
+  ) where
 
 import           Control.Monad                              (return, (>>))
 import           Data.Binary                                (Binary (..))
 import           Data.Bool                                  ((&&))
-import           Data.Eq                                    (Eq (..))
 import           Data.Function                              (($))
-import           Prelude                                    (Integer)
+import           Prelude                                    (Bool, Integer)
+import qualified Prelude
 import           Text.Show                                  (Show)
 
 import           ZkFold.Base.Algebra.Basic.Class
@@ -29,6 +31,7 @@ import           ZkFold.Base.Algebra.Basic.Number
 import           ZkFold.Base.Algebra.EllipticCurve.Class
 import           ZkFold.Base.Algebra.EllipticCurve.Pairing
 import           ZkFold.Base.Algebra.Polynomials.Univariate (toPoly)
+import           ZkFold.Symbolic.Data.Eq
 
 -------------------------- Scalar field & field towers -------------------------
 
@@ -65,44 +68,44 @@ type Fp12 = Ext2 Fp6 "IP3"
 
 ------------------------------- bn254 G1 ---------------------------------------
 
-data BN254_G1
+type BN254_G1_Point = BN254_G1_PointOf Fp
 
-instance EllipticCurve BN254_G1 where
-  type ScalarField BN254_G1 = Fr
-  type BaseField BN254_G1 = Fp
-  pointGen = pointXY 1 2
-  add = addPoints
-  mul = pointMul
+type BN254_G1_PointOf = Weierstrass "BN254_G1" (Point Bool)
 
-instance WeierstrassCurve BN254_G1 where
-  weierstrassA = 0
-  weierstrassB = 3
+instance Field field => WeierstrassCurve "BN254_G1" field where
+  weierstrassB = fromConstant (3 :: Natural)
+
+instance SubgroupCurve "BN254_G1" Bool Fp Fr BN254_G1_PointOf where
+  pointGen = pointXY one (fromConstant (2 :: Natural))
+
+instance Scale Fr BN254_G1_Point where
+  scale n x = scale (toConstant n) x
 
 ------------------------------- bn254 G2 ---------------------------------------
 
-data BN254_G2
+type BN254_G2_Point = BN254_G2_PointOf Fp2
 
-instance EllipticCurve BN254_G2 where
-  type ScalarField BN254_G2 = Fr
-  type BaseField BN254_G2 = Fp2
-  pointGen = pointXY
-    (Ext2 0x1800deef121f1e76426a00665e5c4479674322d4f75edadd46debd5cd992f6ed
-          0x198e9393920d483a7260bfb731fb5d25f1aa493335a9e71297e485b7aef312c2)
-    (Ext2 0x12c85ea5db8c6deb4aab71808dcb408fe3d1e7690c43d37b4ce6cc0166fa7daa
-          0x90689d0585ff075ec9e99ad690c3395bc4b313370b38ef355acdadcd122975b)
-  add = addPoints
-  mul = pointMul
+type BN254_G2_PointOf = Weierstrass "BN254_G2" (Point Bool)
 
-instance WeierstrassCurve BN254_G2 where
-  weierstrassA = zero
+instance WeierstrassCurve "BN254_G2" Fp2 where
   weierstrassB =
     Ext2 0x2b149d40ceb8aaae81be18991be06ac3b5b4c5e559dbefa33267e6dc24a138e5
          0x9713b03af0fed4cd2cafadeed8fdf4a74fa084e52d1852e4a2bd0685c315d2
 
+instance SubgroupCurve "BN254_G2" Bool Fp2 Fr BN254_G2_PointOf where
+    pointGen = pointXY
+      (Ext2 0x1800deef121f1e76426a00665e5c4479674322d4f75edadd46debd5cd992f6ed
+            0x198e9393920d483a7260bfb731fb5d25f1aa493335a9e71297e485b7aef312c2)
+      (Ext2 0x12c85ea5db8c6deb4aab71808dcb408fe3d1e7690c43d37b4ce6cc0166fa7daa
+            0x90689d0585ff075ec9e99ad690c3395bc4b313370b38ef355acdadcd122975b)
+
+instance Scale Fr BN254_G2_Point where
+  scale n x = scale (toConstant n) x
+
 ------------------------------- Pairing ----------------------------------------
 
 newtype BN254_GT = BN254_GT Fp12
-  deriving (Eq, Show, MultiplicativeSemigroup, MultiplicativeMonoid)
+  deriving (Prelude.Eq, Show, MultiplicativeSemigroup, MultiplicativeMonoid, Eq Prelude.Bool)
 
 instance Exponent BN254_GT Natural where
   BN254_GT e ^ p = BN254_GT (e ^ p)
@@ -115,11 +118,10 @@ deriving via (NonZero Fp12) instance MultiplicativeGroup BN254_GT
 instance Finite BN254_GT where
   type Order BN254_GT = BN254_Scalar
 
-instance Pairing BN254_G1 BN254_G2 where
-  type TargetGroup BN254_G1 BN254_G2 = BN254_GT
+instance Pairing Fr BN254_G1_Point BN254_G2_Point BN254_GT where
   pairing p q
     = BN254_GT
-    $ finalExponentiation @BN254_G2
+    $ finalExponentiation @Fr
     $ millerAlgorithmBN xi param p q
     where
       param = [ 1
@@ -131,9 +133,9 @@ instance Pairing BN254_G1 BN254_G2 where
 
 ------------------------------ Encoding ----------------------------------------
 
-instance Binary (Point BN254_G1) where
-  put (Point xp yp isInf) =
-    if isInf then put @(Point BN254_G1) (pointXY zero zero) else put xp >> put yp
+instance Binary BN254_G1_Point where
+  put (Weierstrass (Point xp yp isInf)) =
+    if isInf then put @BN254_G1_Point (pointXY zero zero) else put xp >> put yp
   get = do
     xp <- get
     yp <- get
@@ -142,9 +144,9 @@ instance Binary (Point BN254_G1) where
       then pointInf
       else pointXY xp yp
 
-instance Binary (Point BN254_G2) where
-  put (Point xp yp isInf) =
-    if isInf then put @(Point BN254_G2) (pointXY zero zero) else put xp >> put yp
+instance Binary BN254_G2_Point where
+  put (Weierstrass (Point xp yp isInf)) =
+    if isInf then put @BN254_G2_Point (pointXY zero zero) else put xp >> put yp
   get = do
     xp <- get
     yp <- get
